move smarts_ros ROS node into smarts/ros so it can be distributed with pkg
