+++ smarts/ros/.catkin_workspace
