+++ smarts/ros/.gitignore
