+++ smarts/ros/src/smarts_ros/scripts/test_smarts_ros.py
